Moves file and indent
